Add: Test for ruby-switch
There are Ruby 2.2 packages in the brightbox repo.  At present they will
be installed, and set to 2.2, which will cause all sorts of problems
because there are no dev packages yet.  Also we don't want to use 2.2.

Ruby-switch is a simple interface to the alternatives system and allows
us to specify that we should use 2.1 even if something else gets installed.

This tests that the tool is installed and that 2.1 is the default.